server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.115.5
fonttools==4.55.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.10.0
mysql-connector-python==9.1.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
pyasn1==0.6.1
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
pytz==2024.2
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
xlrd==2.0.1
